+++ buildSrc/src/main/java/com/google/firebase/gradle/plugins/publish/PublishingPlugin.java
@@ -14,6 +14,7 @@
 
 package com.google.firebase.gradle.plugins.publish;
 
+import com.google.firebase.gradle.plugins.CheckHeadDependencies;
 import com.google.firebase.gradle.plugins.FirebaseLibraryExtension;
 import java.io.File;
 import java.io.IOException;
@@ -117,6 +118,13 @@ public class PublishingPlugin implements Plugin<Project> {
                     List.of(projectNamesToPublish.split(projectsToPublishSeparator, -1));
               }
 
+              Set<String> allFirebaseProjects =
+                  project.getSubprojects().stream()
+                      .map(sub -> sub.getExtensions().findByType(FirebaseLibraryExtension.class))
+                      .filter(ext -> ext != null)
+                      .map(ext -> ext.artifactId.get())
+                      .collect(Collectors.toSet());
+
               Set<FirebaseLibraryExtension> projectsToPublish =
                   projectsNames.stream()
                       .filter(name -> !name.isEmpty())
@@ -125,7 +133,8 @@ public class PublishingPlugin implements Plugin<Project> {
                               project
                                   .project(name)
                                   .getExtensions()
-                                  .getByType(FirebaseLibraryExtension.class))
+                                  .findByType(FirebaseLibraryExtension.class))
+                      .filter(ext -> ext != null)
                       .flatMap(lib -> lib.getLibrariesToRelease().stream())
                       .collect(Collectors.toSet());
               project
@@ -181,6 +190,16 @@ public class PublishingPlugin implements Plugin<Project> {
                           t.dependsOn(getPublishTask(toPublish, "MavenLocal"));
                         }
                       });
+              project
+                  .getTasks()
+                  .create(
+                      "checkHeadDependencies",
+                      CheckHeadDependencies.class,
+                      t -> {
+                        t.getProjectsToPublish().set(projectsToPublish);
+                        t.getAllFirebaseProjects().set(allFirebaseProjects);
+                      });
+
               Task publishProjectsToBuildDir =
                   project
                       .getTasks()